+++ InstallomatorLabels.sh
@@ -1011,7 +1011,8 @@ omnioutliner5)
 omniplan3)
     name="OmniPlan"
     type="dmg"
-    downloadURL=$(curl -fs "https://update.omnigroup.com/appcast/com.omnigroup.OmniPlan3" | xpath '//rss/channel/item[1]/enclosure[1]/@url' 2>/dev/null | head -1 | cut -d '"' -f 2)
+    #downloadURL=$(curl -fs "https://update.omnigroup.com/appcast/com.omnigroup.OmniPlan3" | xpath '//rss/channel/item[1]/enclosure[1]/@url' 2>/dev/null | head -1 | cut -d '"' -f 2)
+    downloadURL=$(curl -fs "https://www.omnigroup.com/download" | grep -i "OmniPlan-3" | head -1 | sed 's/.*href="//' | sed 's/".*//')
     appNewVersion=$( echo "${downloadURL}" | sed -E 's/.*\/[a-zA-Z]*-([0-9.]*)\..*/\1/g' )
     expectedTeamID="34YW5XSRB7"
     ;;
@@ -1321,14 +1322,6 @@ sketch)
      appNewVersion=$(curl -fs https://www.sketch.com/updates/ | grep "Sketch Version" | head -1 | sed -E 's/.*Version ([0-9.]*)<.*/\1/g') # version from update page
      expectedTeamID="WUGMZZ5K46"
      ;;
-skitch)
-    # credit: Isaac Ordonez, Mann consulting (@mannconsulting)
-    name="Skitch"
-    type="zip"
-    downloadURL=$(curl -fs -A "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_14_4) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/12.1 Safari/605.1.15" https://evernote.com/products/skitch | grep -o "https://.*zip")
-    expectedTeamID="J8RPQ294UB"
-    Company="Evernote"
-    ;;
 skype)
     name="Skype"
     type="dmg"